gst/mpegstream/: and also fix integer overflows at high PTS values (see #126967).
Original commit message from CVS:
2004-01-03  Ronald Bultje  <rbultje@ronald.bitfreak.net>

* gst/mpegstream/gstmpegdemux.c: (gst_mpeg_demux_parse_pes):
* gst/mpegstream/gstmpegparse.c: (gst_mpeg_parse_class_init),
(gst_mpeg_parse_init):
* gst/mpegstream/gstmpegparse.h:
Remove clock (which was never provided, i.e. dead code), and
also fix integer overflows at high PTS values (see #126967).

+++ gst/mpegstream/gstmpegdemux.c
@@ -746,14 +746,14 @@ gst_mpeg_demux_parse_pes (GstMPEGParse *mpeg_parse, GstBuffer *buffer)
     /* check for PTS */
     if ((flags2 & 0x80)) {
     /*if ((flags2 & 0x80) && id == 0xe0) { */
-      pts  = (*buf++ & 0x0E) << 29;
-      pts |=  *buf++         << 22;
-      pts |= (*buf++ & 0xFE) << 14;
-      pts |=  *buf++         <<  7;
-      pts |= (*buf++ & 0xFE) >>  1;
+      pts  = ((guint64) (*buf++ & 0x0E)) << 29;
+      pts |= ((guint64)  *buf++        ) << 22;
+      pts |= ((guint64) (*buf++ & 0xFE)) << 14;
+      pts |= ((guint64)  *buf++        ) <<  7;
+      pts |= ((guint64) (*buf++ & 0xFE)) >>  1;
 
       GST_DEBUG ("%x PTS = %" G_GUINT64_FORMAT, 
-		      id, MPEGTIME_TO_GSTTIME (pts));
+		 id, MPEGTIME_TO_GSTTIME (pts));
 
     }
     if ((flags2 & 0x40)) {
